+++ src/TextArea/TextArea.js
@@ -4,6 +4,10 @@ import Label from '../Label';
 import InputArea from '../InputArea';
 import InputAreaWithLabelComposite from '../Composite/InputAreaWithLabelComposite/InputAreaWithLabelComposite';
 
+import deprecationLog from '../utils/deprecationLog';
+
+deprecationLog(`Using "TextArea" is deprecated. Please see "3.2a Text Area" updated documentation for a composite usage of "FormField" and "TextInput".`);
+
 const TextArea = ({children, ...props}) => (
   <InputAreaWithLabelComposite {...props}>
     {children}