+++ stories/BadgeSelect/index.story.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import BadgeSelect from 'wix-style-react/BadgeSelect';
 import {SKIN, TYPE, SIZE} from 'wix-ui-backoffice/dist/src/components/Badge/constants';
 import {storySettings} from './storySettings';
@@ -8,11 +9,26 @@ const options = Object.values(SKIN).map((skin, id) => ({
   text: skin
 }));
 
+// Centering the component since the DropdownLayout is centerized and overflows
+// the AutoExample container. In eyes test, the overflowed part will be cut off.
+const CenterBadgeSelect = props => (
+  <div
+    style={{
+      display: 'flex',
+      justifyContent: 'center'
+    }}
+    >
+    <BadgeSelect {...props}/>
+  </div>
+);
+
+CenterBadgeSelect.displayName = BadgeSelect.displayName;
+
 export default {
   category: storySettings.kind,
   storyName: storySettings.storyName,
 
-  component: BadgeSelect,
+  component: CenterBadgeSelect,
   componentPath: '../../src/BadgeSelect',
 
   componentProps: setState => ({