+++ stories/TextArea/storySettings.js
@@ -1,0 +1,6 @@
+import {Category} from '../storiesHierarchy';
+
+export const storySettings = {
+  category: Category.INPUTS,
+  storyName: '3.2a Text Area'
+};