+++ stories/Tooltip/content.js
@@ -1,0 +1,2 @@
+export const SHORT_CONTENT = 'Hello World';
+export const LONG_CONTENT = 'Hello World, Hello World, Hello World, Hello World';